feat(cli): update build script to create Mould Components
