+++ cli/constants.js
@@ -1,9 +1,11 @@
 'use strict'
 
+const COMPONENTS_DIRECTORY = '.components'
 const MOULD_DIRECTORY = 'mould'
 const RESOLVERS = 'resolvers.ts'
 const SYMLINK_MOULD_DIRECTORY = '.mould'
 
+exports.COMPONENTS_DIRECTORY = COMPONENTS_DIRECTORY
 exports.MOULD_DIRECTORY = MOULD_DIRECTORY
 exports.RESOLVERS = RESOLVERS
 exports.SYMLINK_MOULD_DIRECTORY = SYMLINK_MOULD_DIRECTORY
