+++ index.ts
@@ -1,9 +1,2 @@
-import Components from './components'
-
-export const components = {}
-export const transforms = {}
-
-Components.forEach((comp) => {
-    components[comp.type] = comp.Raw
-    transforms[comp.type] = comp.Transform
-})
+export * from './.components'
+export * from './mould'
